requirements.txt
absl-py==2.1.0
attrs==24.2.0
Flask==3.0.3
Flask-Cors==4.0.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
numpy==1.26.4
Pillow==10.4.0
python-dotenv==1.0.1
tensorflow==2.10.1
keras==2.10.0
absl-py==2.1.0
Werkzeug==3.0.3
gunicorn==20.0.4
statsmodels==0.14.0
